requirements.txt
# Core Web Framework
Flask==3.0.0
Werkzeug==2.3.7
gunicorn==21.2.0
gevent==23.9.1

# Database
psycopg2-binary==2.9.10  # PostgreSQL adapter
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1

# Security and Authentication
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
PyJWT==2.8.0
bcrypt==4.0.1

# Session Management
Flask-Session==0.5.0
redis==5.0.1

# HTTP and Networking
requests==2.31.0
urllib3==2.0.7

# API and Data Handling
Flask-RESTful==0.3.10
marshmallow==3.20.1
python-dotenv==1.0.0

# File Handling and Processing
Pillow==10.1.0  # For image processing
python-magic==0.4.27

# Email Support
Flask-Mail==0.9.1

# Monitoring and Logging
sentry-sdk==1.32.0
Flask-MonitoringDashboard==3.1.2

# Development Tools
python-dateutil==2.8.2
itsdangerous==2.1.2
six==1.16.0

# Testing
pytest==7.4.2
pytest-cov==4.1.0
